hybrid_agentic_system/requirements.txt
# Core
fastapi==0.110.1
uvicorn[standard]==0.29.0
google-generativeai==0.3.2

# LangGraph (latest working)
langgraph==0.0.31

# Compatible LangChain + Core
langchain==0.1.13
langchain-core==0.1.52
langsmith==0.1.20
langchain-community==0.0.35
langchain-text-splitters==0.0.1

# Optional
python-dotenv==1.0.1
pydantic==2.6.4